+++ learning_credentials/models.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import json
 import logging
-import uuid
+import uuid as uuid_lib
 from importlib import import_module
 from pathlib import Path
 from typing import TYPE_CHECKING, Self
@@ -16,6 +16,7 @@ from django.core.exceptions import ValidationError
 from django.db import models
 from django.db.models.signals import post_delete
 from django.dispatch import receiver
+from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 from django_celery_beat.models import IntervalSchedule, PeriodicTask
 from edx_ace import Message, Recipient, ace
@@ -25,7 +26,7 @@ from opaque_keys.edx.django.models import LearningContextKeyField
 from learning_credentials.compat import get_learning_context_name
 from learning_credentials.exceptions import AssetNotFoundError, CredentialGenerationError
 
-if TYPE_CHECKING:  # pragma: no cover
+if TYPE_CHECKING:
     from django.core.files import File
     from django.db.models import QuerySet
 
@@ -175,8 +176,7 @@ class CredentialConfiguration(TimeStampedModel):
                  2. Have such a credential with an error status.
         """
         users_ids_with_credentials = Credential.objects.filter(
-            models.Q(learning_context_key=self.learning_context_key),
-            models.Q(credential_type=self.credential_type),
+            models.Q(configuration=self),
             ~(models.Q(status=Credential.Status.ERROR)),
         ).values_list('user_id', flat=True)
 
@@ -197,7 +197,7 @@ class CredentialConfiguration(TimeStampedModel):
         custom_options = _deep_merge(self.credential_type.custom_options, self.custom_options)
         return func(self.learning_context_key, custom_options)
 
-    def generate_credential_for_user(self, user_id: int, celery_task_id: int = 0):
+    def generate_credential_for_user(self, user_id: int, celery_task_id: int = 0) -> Credential:
         """
         Celery task for processing a single user's credential.
 
@@ -208,19 +208,26 @@ class CredentialConfiguration(TimeStampedModel):
         Args:
             user_id: The ID of the user to process the credential for.
             celery_task_id (optional): The ID of the Celery task that is running this function.
+
+        Returns:
+            The generated Credential object.
         """
         user = get_user_model().objects.get(id=user_id)
         # Use the name from the profile if it is not empty. Otherwise, use the first and last name.
         # We check if the profile exists because it may not exist in some cases (e.g., when a User is created manually).
         user_full_name = getattr(getattr(user, 'profile', None), 'name', f"{user.first_name} {user.last_name}")
+        learning_context_name = get_learning_context_name(self.learning_context_key)
         custom_options = _deep_merge(self.credential_type.custom_options, self.custom_options)
 
-        credential, _ = Credential.objects.update_or_create(
-            user_id=user_id,
+        credential, _ = Credential.objects.exclude(status=Credential.Status.INVALIDATED).update_or_create(
+            user=user,
+            configuration=self,
+            # TODO: Remove learning_context_key and credential_type after removing them from the Credential model.
             learning_context_key=self.learning_context_key,
             credential_type=self.credential_type.name,
             defaults={
                 'user_full_name': user_full_name,
+                'learning_context_name': learning_context_name,
                 'status': Credential.Status.GENERATING,
                 'generation_task_id': celery_task_id,
             },
@@ -232,19 +239,21 @@ class CredentialConfiguration(TimeStampedModel):
             generation_func = getattr(generation_module, generation_func_name)
 
             # Run the functions. We do not validate them here, as they are validated in the model's clean() method.
-            credential.download_url = generation_func(self.learning_context_key, user, credential.uuid, custom_options)
+            credential.download_url = generation_func(credential, custom_options)
             credential.status = Credential.Status.AVAILABLE
             credential.save()
         except Exception as exc:
             credential.status = Credential.Status.ERROR
             credential.save()
-            msg = f'Failed to generate the {credential.uuid=} for {user_id=} with {self.id=}.'
+            msg = f'Failed to generate the {credential.uuid=} for {user_id=} with {self.id=}.\nReason: {exc}'
             raise CredentialGenerationError(msg) from exc
         else:
             # TODO: In the future, we want to check this before generating the credential.
             #       Perhaps we could even include this in a processor to optimize it.
             if user.is_active and user.has_usable_password():
                 credential.send_email()
+
+        return credential
 
 
 # noinspection PyUnusedLocal
@@ -280,17 +289,44 @@ class Credential(TimeStampedModel):
 
     uuid = models.UUIDField(
         primary_key=True,
-        default=uuid.uuid4,
+        default=uuid_lib.uuid4,
         editable=False,
         help_text=_('Auto-generated UUID of the credential'),
     )
-    user_id = models.IntegerField(help_text=_('ID of the user receiving the credential'))
-    user_full_name = models.CharField(max_length=255, help_text=_('User receiving the credential'))
+    verify_uuid = models.UUIDField(
+        default=uuid_lib.uuid4,
+        editable=False,
+        help_text=_('UUID used for verifying the credential'),
+    )
+    user = models.ForeignKey(
+        settings.AUTH_USER_MODEL,
+        on_delete=models.CASCADE,
+        help_text=_('User receiving the credential'),
+    )
+    user_full_name = models.CharField(
+        max_length=255,
+        editable=False,
+        help_text=_('User receiving the credential. This field is used for validation purposes.'),
+    )
+    # TODO: Remove credential_type and learning_context_key. They are redundant because we have the configuration FK.
     learning_context_key = LearningContextKeyField(
         max_length=255,
         help_text=_('ID of a learning context (e.g., a course or a Learning Path) for which the credential was issued'),
     )
+    learning_context_name = models.CharField(
+        max_length=255,
+        editable=False,
+        help_text=_(
+            'Name of the learning context for which the credential was issued. '
+            'This field is used for validation purposes.'
+        ),
+    )
     credential_type = models.CharField(max_length=255, help_text=_('Type of the credential'))
+    configuration = models.ForeignKey(
+        'CredentialConfiguration',
+        on_delete=models.PROTECT,
+        help_text=_('Associated credential configuration'),
+    )
     status = models.CharField(
         max_length=32,
         choices=Status.choices,
@@ -300,29 +336,60 @@ class Credential(TimeStampedModel):
     download_url = models.URLField(blank=True, help_text=_('URL of the generated credential PDF (e.g., to S3)'))
     legacy_id = models.IntegerField(null=True, help_text=_('Legacy ID of the credential imported from another system'))
     generation_task_id = models.CharField(max_length=255, help_text=_('Task ID from the Celery queue'))
+    invalidated_at = models.DateTimeField(
+        null=True, editable=False, help_text=_('Timestamp when the credential was invalidated')
+    )
+    invalidation_reason = models.CharField(
+        max_length=255, blank=True, help_text=_('Reason for invalidating the credential')
+    )
 
-    class Meta:  # noqa: D106
-        unique_together = (('user_id', 'learning_context_key', 'credential_type'),)
+    def __str__(self):
+        """Get a string representation of this model's instance."""
+        return (
+            f"{self.configuration.credential_type.name} for {self.user_full_name} "
+            f"in {self.configuration.learning_context_key}"
+        )
 
-    def __str__(self):  # noqa: D105
-        return f"{self.credential_type} for {self.user_full_name} in {self.learning_context_key}"
+    def save(self, *args, **kwargs):
+        """If the invalidation reason is set, trigger the invalidation and update the timestamp."""
+        if self.invalidation_reason and self.status != Credential.Status.INVALIDATED:
+            self._invalidate()
+        if self.status == Credential.Status.INVALIDATED and not self.invalidated_at:
+            self.invalidated_at = timezone.now()
+        super().save(*args, **kwargs)
+
+    def _invalidate(self):
+        """Trigger the invalidation process for the credential."""
+        generation_module_name, generation_func_name = self.configuration.credential_type.generation_func.rsplit('.', 1)
+        generation_module = import_module(generation_module_name)
+        generation_func = getattr(generation_module, generation_func_name)
+
+        self.download_url = generation_func(self, {}, invalidate=True)
+        self.status = Credential.Status.INVALIDATED
 
     def send_email(self):
         """Send a credential link to the student."""
-        learning_context_name = get_learning_context_name(self.learning_context_key)
-        user = get_user_model().objects.get(id=self.user_id)
         msg = Message(
             name="certificate_generated",
             app_label="learning_credentials",
-            recipient=Recipient(lms_user_id=user.id, email_address=user.email),
+            recipient=Recipient(lms_user_id=self.user.id, email_address=self.user.email),
             language='en',
             context={
                 'certificate_link': self.download_url,
-                'course_name': learning_context_name,
+                'course_name': self.learning_context_name,
                 'platform_name': settings.PLATFORM_NAME,
             },
         )
         ace.send(msg)
+
+    def reissue(self) -> Self:
+        """Invalidate the current credential and create a new one."""
+        if self.invalidation_reason:
+            self.invalidation_reason += '\n'
+        self.invalidation_reason += 'Reissued'
+        self.save()
+
+        return self.configuration.generate_credential_for_user(self.user.id)
 
 
 class CredentialAsset(TimeStampedModel):